% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{figureFilename}
\alias{figureFilename}
\title{Adds a timestamp to a filename}
\usage{
figureFilename(title)
}
\arguments{
\item{title}{The filename to output}
}
\description{
Quick way to add one of Sean's Preferred Timestamps (TM) to a figure filename.
}
\keyword{date}
\keyword{time}

