% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-msigdb.R
\name{get_msigdb_gene_set}
\alias{get_msigdb_gene_set}
\title{Search MSigDB For a Given Gene Set}
\usage{
get_msigdb_gene_set(gene.set.name)
}
\arguments{
\item{gene.set.name}{The name (in whole, or part!) of the gene set for which to find gene IDs}
}
\value{
The MSigDB gene set's Entrez IDs, as a character vector. Or, a list of gene sets, if there were multiple matches. The leading part of each result gene set (the portion before the '.') indicates which gene set collection the gene set was found in.
}
\description{
This is a convenience function to look through the MSigDB gene set collections (C1-C7 and Hallmark) for a gene set
with a known name, and to return the Entrez gene IDs associated with that gene set. Pretty simple!
}
\keyword{msigdb}

