% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{notebookTimestamp}
\alias{notebookTimestamp}
\title{Makes a timestamp in the format Sean likes}
\usage{
notebookTimestamp()
}
\description{
In all the figures and files I make during analysis, I add a timestamp prefix of the 
format "YYMMDD". That way, I can sort by filename and have them in chronolgical order,
and not have to worry about the timestamps created by the filesystem.
}
\keyword{date}
\keyword{time}

