% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-modeling.R
\name{run.feature.lm}
\alias{run.feature.lm}
\title{Runs a linear model for a single given phenotypic feature}
\usage{
run.feature.lm(feature.name, pheno.data, dataset, is.numeric = TRUE)
}
\arguments{
\item{feature.name}{A string of the column name in the phenotype data matrix corresponding to the feature of interest}

\item{pheno.data}{The full phenotypic data matrix (e.g. pData(expression.set))}

\item{dataset}{the expression matrix of interest. ncol(dataset) should == nrow(pheno.data)}

\item{is.numeric}{Whether the feature in question is numeric or not. If not numeric, it's assumed to be a factor.}
}
\description{
This function is a wrapper for limma::lmFit, which is usually used to run linear models for microarray data.
It returns a dataframe with the t-statistic, p-value, and FDR adjusted p-value for each gene.
}
\keyword{clinical}
\keyword{feature}
\keyword{limma}
\keyword{linear}
\keyword{microarray}
\keyword{model}

